Deprecate testing classes in org.apache.maven.api, add new ones in org.apache.maven.testing
Create new testing classes in org.apache.maven.testing.di and
org.apache.maven.testing.plugin packages. Deprecate the old classes
in org.apache.maven.api.di.testing and org.apache.maven.api.plugin.testing
for removal in a future version.
