+++ impl/maven-testing/src/main/java/org/apache/maven/testing/di/MavenDITest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.di.testing;
+package org.apache.maven.testing.di;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -26,9 +26,45 @@ import java.lang.annotation.Target;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
- * @deprecated Use {@link org.apache.maven.testing.di.MavenDITest} instead
+ * Annotation that enables Maven's dependency injection support in JUnit tests.
+ * When applied to a test class, it automatically sets up the DI container and
+ * performs injection into test instances.
+ *
+ * <p>This annotation is a convenient way to use {@link MavenDIExtension} without
+ * explicitly using {@code @ExtendWith}. It provides the same functionality as the
+ * legacy Plexus test support but uses Maven's new DI framework.</p>
+ *
+ * <p>Example usage:</p>
+ * <pre>
+ * {@code
+ * @MavenDITest
+ * class MyComponentTest {
+ *     @Inject
+ *     private MyComponent component;
+ *
+ *     @Test
+ *     void testComponentBehavior() {
+ *         // component is automatically injected
+ *         assertNotNull(component);
+ *         // perform test
+ *     }
+ * }
+ * }
+ * </pre>
+ *
+ * <p>The annotation supports:</p>
+ * <ul>
+ *   <li>Constructor injection</li>
+ *   <li>Field injection</li>
+ *   <li>Method injection</li>
+ *   <li>Automatic component discovery</li>
+ *   <li>Lifecycle management of injected components</li>
+ * </ul>
+ *
+ * @see MavenDIExtension
+ * @see org.apache.maven.api.di.Inject
+ * @since 4.0.0
  */
-@Deprecated(since = "4.0.0-rc-6", forRemoval = true)
 @Retention(RetentionPolicy.RUNTIME)
 @ExtendWith(MavenDIExtension.class)
 @Target(ElementType.TYPE)
