+++ impl/maven-testing/src/main/java/org/apache/maven/testing/plugin/Basedir.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.plugin.testing;
+package org.apache.maven.testing.plugin;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -24,8 +24,51 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** @deprecated Use {@link org.apache.maven.testing.plugin.Basedir} instead */
-@Deprecated(since = "4.0.0-rc-6", forRemoval = true)
+/**
+ * Specifies the base directory for test resources in Maven plugin tests.
+ * This annotation can be applied to test classes or methods to define where test resources
+ * (such as POM files, source files, and other test artifacts) are located.
+ *
+ * <p>If not specified, the plugin's base directory will be used as the default.</p>
+ *
+ * <p>Example usage on class level:</p>
+ * <pre>
+ * {@code
+ * @MojoTest
+ * @Basedir("src/test/resources/my-test-project")
+ * class MyMojoTest {
+ *     @Test
+ *     @InjectMojo(goal = "compile")
+ *     void testCompilation(MyMojo mojo) {
+ *         // Test resources will be loaded from src/test/resources/my-test-project
+ *         mojo.execute();
+ *     }
+ * }
+ * }
+ * </pre>
+ *
+ * <p>Example usage on method level:</p>
+ * <pre>
+ * {@code
+ * @MojoTest
+ * class MyMojoTest {
+ *     @Test
+ *     @Basedir("src/test/resources/specific-test-case")
+ *     @InjectMojo(goal = "compile")
+ *     void testSpecificCase(MyMojo mojo) {
+ *         // Test resources will be loaded from src/test/resources/specific-test-case
+ *         mojo.execute();
+ *     }
+ * }
+ * }
+ * </pre>
+ *
+ * <p>When applied at both class and method level, the method-level annotation takes precedence.</p>
+ *
+ * @see MojoTest
+ * @see MojoExtension
+ * @since 4.0.0
+ */
 @Target({ElementType.TYPE, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
