+++ impl/maven-testing/src/main/java/org/apache/maven/testing/plugin/MojoParameters.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.maven.api.plugin.testing;
+package org.apache.maven.testing.plugin;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Inherited;
@@ -24,8 +24,50 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/** @deprecated Use {@link org.apache.maven.testing.plugin.MojoParameters} instead */
-@Deprecated(since = "4.0.0-rc-6", forRemoval = true)
+/**
+ * Container annotation for multiple {@link MojoParameter} annotations.
+ * This annotation is automatically used by Java when multiple {@code @MojoParameter}
+ * annotations are applied to the same element.
+ *
+ * <p>While this annotation can be used directly, it's generally more convenient
+ * to use multiple {@code @MojoParameter} annotations, which Java will automatically
+ * wrap in this container annotation.</p>
+ *
+ * <p>Example of direct usage:</p>
+ * <pre>
+ * {@code
+ * @Test
+ * @InjectMojo(goal = "compile")
+ * @MojoParameters({
+ *     @MojoParameter(name = "source", value = "1.8"),
+ *     @MojoParameter(name = "target", value = "1.8"),
+ *     @MojoParameter(name = "debug", value = "true")
+ * })
+ * void testCompilation(CompileMojo mojo) {
+ *     mojo.execute();
+ * }
+ * }
+ * </pre>
+ *
+ * <p>Equivalent usage with repeatable annotation:</p>
+ * <pre>
+ * {@code
+ * @Test
+ * @InjectMojo(goal = "compile")
+ * @MojoParameter(name = "source", value = "1.8")
+ * @MojoParameter(name = "target", value = "1.8")
+ * @MojoParameter(name = "debug", value = "true")
+ * void testCompilation(CompileMojo mojo) {
+ *     mojo.execute();
+ * }
+ * }
+ * </pre>
+ *
+ * @see MojoParameter
+ * @see InjectMojo
+ * @see MojoTest
+ * @since 4.0.0
+ */
 @Target({ElementType.METHOD, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
 @Inherited
