+++ impl/maven-testing/src/test/java/org/apache/maven/api/plugin/testing/MojoRealSessionTest.java
@@ -25,8 +25,9 @@ import org.apache.maven.api.Session;
 import org.apache.maven.api.di.Inject;
 import org.apache.maven.api.di.Provides;
 import org.apache.maven.api.di.Singleton;
-import org.apache.maven.api.plugin.testing.stubs.SessionMock;
 import org.apache.maven.impl.standalone.ApiRunner;
+import org.apache.maven.testing.plugin.MojoTest;
+import org.apache.maven.testing.plugin.stubs.SessionMock;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
